♻️ [Refactor] ScheduleVM -> HomVM으로 갈음
